requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
authlib>=1.3.0
python-dotenv>=1.0.0
python-docx>=1.1.0
jinja2>=3.1.0
pdfminer.six>=20250324
PyMuPDF>=1.24.0
xhtml2pdf>=0.2.17
reportlab>=4.0.4
